requirements.txt
# Requirements for Flask
Flask==2.2.2
WTForms==3.0.1
Flask-WTF==1.0.1
Flask-MySQL==1.5.2
itsdangerous==2.1.2
jinja2==3.1.2
Werkzeug==2.2.2
click==8.1.3
MarkupSafe==2.1.2
python-dateutil==2.8.2
urllib3==1.26.14
idna==3.4
certifi==2022.12.7
charset-normalizer==3.0.1
six==1.16.0

python-dotenv==0.21.0	# dotenv
python-crontab==2.7.1	# cron
requests==2.28.2		# requests
python-scrapyd-api==2.1.2		# scrapyd

networkx~=3.1
pyvis~=0.3.2